DSEGOG-179 Rename filename to correct typo
